Major refactor. GameController: Is now a proper control class that acts as the middle man between the model and view. GameController is now also responsible for constructing the games GameState object.
GameState:
ActorList and Player are no longer static

Main:
Updated the main method calls to reflect refactor

Factory's now contain a reference to the GameController, but I dont think this is a final solution as the reference is only needed to pass the GameKeyListener to Player and potentially ActorStrategy.

Player:
Now has a reference to GameController that it uses to check for collisions instead of referring the static GameState method.

DataStorage:
Now takes a reference to GameController and use's its getter to for the gameState before writing to file