Login Screen complete.
Changes:
MainServer and MainClient are now objects that get instantiated when
user clicks on either Host or Join in Login Screen. 
They then make the call to launchServer()/launchClient().

To run the game the user must run the main in LoginGame. 

Inventory working again. 

Made the client have a getter for the gameState again - Not sure if
this is good practice, but it works and doesn't effect game performance. 

Uncommented the GameCanvas if statement which gets thie player and
updates their inventory via the above getter.